+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v1.8.0 > vx.x.x
 
+-   Add new publication services: Rumble, PicartoTV, NimoTV, Livepush
 -   Add frame interpolation (framerate) filter (thanks to orryverducci)
 -   Add -referer option for pulling HTTP streams ([PR 40](https://github.com/datarhei/restreamer-ui/pull/40), thanks to mdastgheib)
 -   Add a/v filter to the publication components ([#593](https://github.com/datarhei/restreamer-ui/issues/593))